// BooklingSystem.UI/src/models/OrderActions.tsx
export enum OrderActions {
  SET_STEP_INC = "SETSTEPINC",
  SET_STEP_DEC = "SETSTEPDEC",
  SET_STEP_1 = "SETSTEP1",
  SET_SIZE = "SETSIZE",
  SET_BBQ = "SETBBQ",
  SET_SPEAKER = "SETSPEAKER",
  SET_CHAIRS = "SETCHAIRS",
  SET_TABLES = "SETTABLES",
  SET_DATE = "SETDATE",
  SET_ADDRESS = "SETADDRESS",
  SET_USER = "SETUSER",
  SET_PRICE = "SETPRICE",
  SET_PRICE2 = "SETPRICE2",
  SET_TENTID = "SET_TENTID",
}